requirements.txt
backtrader
pandas
numpy
yfinance
python-binance
python-telegram-bot
PyYAML
streamlit
pytest
scikit-optimize
SQLAlchemy
matplotlib
plotly
requests
ta